refactor: start moving tasks to operations dir
